amule.xpm: recover it back, fix Issue #6.
In commit 2312eccd45a6eee2676e5acdb9e3d2fff923ae66, amule.xpm
was deleted as a "cleanup", however, this file is still required
for installation and listed in Makefile.am, as:

    if MONOLITHIC
    dist_icon_DATA += amule.xpm
    dist_util_DATA += amule.desktop
    endif

The missing amule.xpm produces a build error at the end:

    make: *** No rule to make target 'amule.xpm', needed by 'all-am'.
    Stop.

and causes Issue #6, at that time, the issue was not understood
and we got it confused with the xpm flag icons.

This commit recovers amule.xpm back and resolves this issue.

Signed-off-by: Tom Li <biergaizi@member.fsf.org>